+++ src/app.module.ts
@@ -4,8 +4,6 @@ import { PrismaModule } from 'prisma/prisma.module';
 import { AdminFileModule } from './core/admin-file';
 import { ResponseInterceptor } from './core/interceptors/response.interceptor';
 import { UsersModule } from './modules/users/users.module';
-import { ClientesModule } from './modules/clientes/clientes.module';
-import { ServiciosModule } from './modules/servicios/servicios.module';
 import { ProcessIfcModule } from './modules/process-ifc/process-ifc.module';
 import { RedisModule } from './core/redis/redis.module';
 import { IfcProcessingModule } from './core/ifc-processing/ifc-processing.module';
@@ -15,8 +13,6 @@ import { IfcProcessingModule } from './core/ifc-processing/ifc-processing.module
     PrismaModule,
     AdminFileModule,
     UsersModule,
-    ClientesModule,
-    ServiciosModule,
     ProcessIfcModule,
     RedisModule,
     IfcProcessingModule