+++ src/core/ifc-processing/interfaces/ifc-processing.interface.ts
@@ -43,6 +43,7 @@ export interface IfcProcessedData {
   categorias: IfcCategoriaResumen[];
   elementosPorNivel: IfcElementosPorNivel[];
   materiales: IfcMaterialResumen[];
+  elementsDB: any;
 }
 
 /**