+++ src/modules/clientes/interfaces/identity-type.interfaces.ts
@@ -1,5 +1,0 @@
-import { IdentityTypeEntity } from '../entities/identity-type.entity';
-
-export abstract class IdentityTypeRepository {
-    abstract findAll(): Promise<IdentityTypeEntity[]>;
-}