+++ src/modules/process-ifc/interfaces/processifc.interfaces.ts
@@ -8,4 +8,6 @@ export type CreateRefFileData = {
 export abstract class ProcessIfcRepository {
   abstract create(data: CreateRefFileData): Promise<RefFileEntity>;
   abstract findAll(): Promise<RefFileEntity[]>;
+  abstract getElementsDBByIfcProcessNivel(id_process: number): Promise<any>;
+  abstract getElementsDBByIfcProcessCategory(id_process: number): Promise<any>;
 }