+++ src/modules/servicios/entities/tipo-servicio.entity.ts
@@ -1,6 +1,0 @@
-export class TipoServicioEntity {
-    constructor(
-        public readonly id: number,
-        public readonly nombre: string,
-    ) { }
-}