+++ src/modules/servicios/interfaces/tipo-servicio.interfaces.ts
@@ -1,5 +1,0 @@
-import { TipoServicioEntity } from '../entities/tipo-servicio.entity';
-
-export abstract class TipoServicioRepository {
-    abstract findAll(): Promise<TipoServicioEntity[]>;
-}